app/requirements.txt
Pillow==4.3.0
requests==2.18.4
RPi.GPIO==0.6.3
spidev==3.2
pyyaml==3.12
